requirements.txt
pandas
numpy
scikit-learn
xgboost
shap
lime
matplotlib
seaborn
textstat
groq
python-dotenv
